Move season and notification logic out of PlantingSiteStore
